requirements.txt
aiohttp==3.9.5 ; python_version >= '3.8'
aiosignal==1.3.2 ; python_version >= '3.9'
annotated-types==0.7.0 ; python_version >= '3.8'
anyio==4.8.0 ; python_version >= '3.9'
attrs==24.3.0 ; python_version >= '3.8'
beautifulsoup4==4.12.3
certifi==2024.12.14 ; python_version >= '3.6'
charset-normalizer==3.4.1 ; python_version >= '3.7'
dataclasses-json==0.6.7 ; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0 ; python_version >= '3.6'
frozenlist==1.5.0 ; python_version >= '3.8'
greenlet==3.1.1 ; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
httpcore==1.0.7 ; python_version >= '3.8'
httpx==0.28.1 ; python_version >= '3.8'
httpx-sse==0.4.0 ; python_version >= '3.8'
idna==3.10 ; python_version >= '3.6'
jiter==0.8.2 ; python_version >= '3.8'
jsonpatch==1.33 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0 ; python_version >= '3.7'
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29 ; python_version >= '3.9' and python_version < '4.0'
langchain-openai==0.2.14
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.4 ; python_version >= '3.9' and python_version < '4.0'
langsmith==0.2.10 ; python_version >= '3.9' and python_version < '4.0'
marshmallow==3.23.3 ; python_version >= '3.9'
multidict==6.1.0 ; python_version >= '3.8'
mypy-extensions==1.0.0 ; python_version >= '3.5'
numpy==1.26.4 ; python_version < '3.12'
openai==1.59.3 ; python_version >= '3.8'
orjson==3.10.13 ; platform_python_implementation != 'PyPy'
packaging==24.2 ; python_version >= '3.8'
pinecone-client==5.0.1 ; python_version >= '3.8' and python_version < '4.0'
pinecone-plugin-inference==1.1.0 ; python_version >= '3.8' and python_version < '4.0'
pinecone-plugin-interface==0.0.7 ; python_version >= '3.8' and python_version < '4.0'
propcache==0.2.1 ; python_version >= '3.9'
pydantic==2.10.4 ; python_version >= '3.8'
pydantic-core==2.27.2 ; python_version >= '3.8'
pydantic-settings==2.7.1 ; python_version >= '3.8'
python-dotenv==1.0.1
pyyaml==6.0.2 ; python_version >= '3.8'
regex==2024.11.6 ; python_version >= '3.8'
requests==2.32.3 ; python_version >= '3.8'
requests-toolbelt==1.0.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1 ; python_version >= '3.7'
soupsieve==2.6 ; python_version >= '3.8'
sqlalchemy==2.0.36 ; python_version >= '3.7'
tenacity==9.0.0 ; python_version >= '3.8'
tiktoken==0.8.0 ; python_version >= '3.9'
tqdm==4.67.1 ; python_version >= '3.7'
typing-extensions==4.12.2 ; python_version >= '3.8'
typing-inspect==0.9.0
urllib3==2.3.0 ; python_version < '3.12' and python_version >= '3.8'
yarl==1.18.3 ; python_version >= '3.9'
streamlit==1.41.1